♻️ Refactor common functions into libs 💄 Code style 📝 Update README
