+++ contracts/scripts/.env.example
@@ -1,0 +1,9 @@
+INFURA_API_KEY=
+
+# STARKNET_NETWORK=sepolia
+# STARKNET_ACCOUNT_ADDRESS=
+# STARKNET_ACCOUNT_PRIVATE_KEY=
+
+STARKNET_NETWORK=goerli
+STARKNET_ACCOUNT_ADDRESS=
+STARKNET_ACCOUNT_PRIVATE_KEY=
