+++ contracts/scripts/package.json
@@ -13,5 +13,8 @@
     "colors": "^1.4.0",
     "dotenv": "^16.3.1",
     "starknet": "^5.24.3"
+  },
+  "devDependencies": {
+    "prettier": "3.1.1"
   }
 }
